+++ app/src/main/java/com/juane/arduino/gpstracker/ui/settings/SettingsViewModel.java
@@ -1,16 +1,16 @@
-package com.juane.arduino.gpstracker.ui.notifications;
+package com.juane.arduino.gpstracker.ui.settings;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class SettingsViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public SettingsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is settings fragment");
     }
 
     public LiveData<String> getText() {
