+++ app/src/main/res/menu/navigation.xml
@@ -9,11 +9,11 @@
     <item
         android:id="@+id/tab2"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/profile_title" />
+        android:title="@string/map_title" />
 
     <item
         android:id="@+id/tab3"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/notifications_title" />
+        android:icon="@drawable/ic_settings_black_24dp"
+        android:title="@string/settings_title" />
 
 </menu>
