+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,6 @@
     <string name="real_time_text">Real Time</string>
     <string name="show_location_text">Show Location</string>
     <string name="home_title">Home</string>
-    <string name="profile_title">Profile</string>
-    <string name="notifications_title">Notifications</string>
+    <string name="map_title">Map</string>
+    <string name="settings_title">Settings</string>
 </resources>
